fix: Derive truncation bound per draw from its sensitivity
Compile epsilon and delta rather than a fixed truncation bound, and derive the
truncated-Laplace bound per released quantity from its L1 sensitivity. A single
compiled bound of 8 only conformed to delta=1/1000 at unit sensitivity. The
impression-threshold draw has sensitivity maxFrequencyPerUser (default 126),
where truncating a scale-126 Laplace at +/-8 left the count effectively unnoised.

- DeterministicTruncatedLaplaceParams: compile EPSILON and DELTA; add
  truncationBound(epsilon, sensitivity) = (sensitivity/epsilon) *
  ln(1 + (e^epsilon - 1)/(2*delta)), via StrictMath for bit-reproducibility.
- ResultNoiser: derive each sampler's bound from its own sensitivity; drop the
  shared truncationBound param and the duplicated unit-sensitivity constant.
- TruncatedLaplaceNoiseDistribution/sampler: truncationBound Int -> Double.
- trus_tee.proto: reserve field 7 (Spanner-persisted ComputationDetails).
- Add TrusTeeProcessorImpl test: deterministic noise ignores the spec privacy
  params. Update goldens (impression draw -4 -> -5 at the corrected bound).

+++ src/main/kotlin/org/wfanet/measurement/computation/DeterministicTruncatedLaplaceNoiseSampler.kt
@@ -31,7 +31,7 @@ class DeterministicTruncatedLaplaceNoiseSampler(
   constructor(
     epsilon: Double,
     sensitivity: Double,
-    truncationBound: Int,
+    truncationBound: Double,
   ) : this(TruncatedLaplaceNoiseDistribution(epsilon, sensitivity, truncationBound))
 
   /**
